Convert the context to ts
