+++ .eslintrc.js
@@ -1,10 +1,11 @@
 module.exports = {
-  extends: '@seasonedsoftware/eslint-config',
-  parser: 'babel-eslint',
+  extends: "@seasonedsoftware/eslint-config",
+  parser: "babel-eslint",
   rules: {
-    'sort-keys': 0,
-    'no-use-before-define': 0,
-    'no-unused-expressions': 0,
-    'import/extensions': 0,
+    "sort-keys": 0,
+    "no-use-before-define": 0,
+    "no-unused-expressions": 0,
+    "import/extensions": 0,
+    "flowtype/no-types-missing-file-annotation": 0,
   },
-}
+};
